backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.9.1
pyjwt==2.9.0